Move aabb and hit to raylib
