+++ src/geom/transform.h
@@ -28,7 +28,6 @@ struct Rotator
 };
 
 // #todo: Generate matrix internally
-// #todo: Support rotation (needs quaternion)
 // #todo: Utilize SIMD
 class Transform
 {
